Add test case from OneLogin
This adds a test case provided by @fredipevcin in [https://github.com/crewjam/saml/pull/25](PR25) that
makes sure we can parse cases where the assertion is signed rather than the whole message. This
has been supported for a while, but it is nice to ensure that the compatibility at issue continues to work.